# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SMRP Backend API description.",
        "title": "Swagger SMRP Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/smrp",
    "paths": {
...
        "/api/discharge-type": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "parameters": [
                    {
                        "description": "Create City Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommonSetupDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/discharge-type/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonSetup"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update City Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommonSetupDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/discharge-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "_page",
                        "name": "_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "_limit",
                        "name": "_limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CommonSetup"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "_page",
                        "name": "_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "_limit",
                        "name": "_limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "description": "Search Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.KeywordDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CommonSetup"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/lookup/discharge-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup/DischargeType"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CommonSetup"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CommonSetupDto": {
            "type": "object",
            "required": [
                "code",
                "desc",
                "ref"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 30
                },
                "desc": {
                    "type": "string",
                    "maxLength": 300
                },
                "ref": {
                    "type": "string",
                    "maxLength": 200
                }
            }
        },
        "dto.KeywordDto": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                }
            }
        },
...
        "model.CommonSetup": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "created_date": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "deleted_date": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modified_by": {
                    "type": "integer"
                },
                "modified_date": {
                    "type": "string"
                },
                "ref": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}